feat: implement logic to get anagrams of two letters
This algorithm is composed of two for loop: the first one will get one by one the letters of the word, fixing that letter into the first position of the combination. The second loop will call again the function with the rest of the word, repeating the same operation until all the letters are done for the fixed letter. This operation will be then repeated to the rest of the letters